feat: User authentication and authorization

+++ routes/api.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Enums\UserRole;
 use App\Http\Controllers\AuthorController;
 use App\Http\Controllers\BookController;
 use App\Http\Controllers\BorrowRecordController;
@@ -7,39 +8,41 @@ use App\Http\Controllers\UserController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
-Route::get('/user', function (Request $request) {
-    return $request->user();
-})->middleware('auth:sanctum');
+$sanctumAUTH = 'auth:sanctum';
+$mustBeAdminOrLibrarian = 'ability:' . UserRole::ADMIN . ',' . UserRole::LIBRARIAN;
+$mustBeAdminOrMember = 'ability:' . UserRole::ADMIN . ',' . UserRole::MEMBER;
+$mustBeAdmin = 'ability:' . UserRole::ADMIN;
+$mustBeMember = 'ability:' . UserRole::MEMBER;
 
-Route::prefix('books')->group(function () {
+Route::prefix('books')->group(function () use ($sanctumAUTH, $mustBeAdminOrLibrarian, $mustBeAdmin, $mustBeMember) {
     Route::get('/', [BookController::class, 'index']);
     Route::get('/{book}', [BookController::class, 'show']);
-    Route::post('/', [BookController::class, 'store']);
-    Route::put('/{book}', [BookController::class, 'update']);
-    Route::delete('/{book}', [BookController::class, 'destroy']);
-    Route::post('/{book}/borrow', [BookController::class, 'borrowBook']);
-    Route::post('/{book}/return', [BookController::class, 'returnBook']);
+    Route::post('/', [BookController::class, 'store'])->middleware($sanctumAUTH, $mustBeAdminOrLibrarian);
+    Route::put('/{book}', [BookController::class, 'update'])->middleware($sanctumAUTH, $mustBeAdminOrLibrarian);
+    Route::delete('/{book}', [BookController::class, 'destroy'])->middleware($sanctumAUTH, $mustBeAdmin);
+    Route::post('/{book}/borrow', [BookController::class, 'borrowBook'])->middleware($sanctumAUTH, $mustBeMember);
+    Route::post('/{book}/return', [BookController::class, 'returnBook'])->middleware($sanctumAUTH, $mustBeMember);
 });
 
-Route::prefix('authors')->group(function () {
+Route::prefix('authors')->group(function () use ($sanctumAUTH, $mustBeAdminOrLibrarian, $mustBeAdmin) {
     Route::get('/', [AuthorController::class, 'index']);
     Route::get('/{author}', [AuthorController::class, 'show']);
-    Route::post('/', [AuthorController::class, 'store']);
-    Route::put('/{author}', [AuthorController::class, 'update']);
-    Route::delete('/{author}', [AuthorController::class, 'destroy']);
+    Route::post('/', [AuthorController::class, 'store'])->middleware($sanctumAUTH, $mustBeAdminOrLibrarian);
+    Route::put('/{author}', [AuthorController::class, 'update'])->middleware($sanctumAUTH, $mustBeAdminOrLibrarian);
+    Route::delete('/{author}', [AuthorController::class, 'destroy'])->middleware($sanctumAUTH, $mustBeAdmin);
 });
 
-Route::prefix('users')->group(function () {
-    Route::get('/', [UserController::class, 'index']);
-    Route::get('/{user}', [UserController::class, 'show']);
+Route::prefix('users')->group(function () use ($sanctumAUTH, $mustBeAdmin, $mustBeAdminOrMember) {
+    Route::get('/', [UserController::class, 'index'])->middleware($sanctumAUTH, $mustBeAdmin);
+    Route::get('/{user}', [UserController::class, 'show'])->middleware($sanctumAUTH, $mustBeAdmin);
     Route::post('/', [UserController::class, 'store']);
-    Route::put('/{user}', [UserController::class, 'update']);
-    Route::delete('/{user}', [UserController::class, 'destroy']);
+    Route::put('/{user}', [UserController::class, 'update'])->middleware($sanctumAUTH, $mustBeAdminOrMember);
+    Route::delete('/{user}', [UserController::class, 'destroy'])->middleware($sanctumAUTH, $mustBeAdmin);
 });
 
 Route::post('/login', [UserController::class, 'login']);
 
-Route::prefix('borrow-records')->group(function () {
-    Route::get('/', [BorrowRecordController::class, 'index']);
-    Route::get('/{borrowRecord}', [BorrowRecordController::class, 'show']);
+Route::prefix('borrow-records')->group(function () use ($sanctumAUTH, $mustBeAdminOrLibrarian) {
+    Route::get('/', [BorrowRecordController::class, 'index'])->middleware($sanctumAUTH, $mustBeAdminOrLibrarian);
+    Route::get('/{borrowRecord}', [BorrowRecordController::class, 'show'])->middleware($sanctumAUTH, $mustBeAdminOrLibrarian);
 });
